+++ HISTORY.md
@@ -1,5 +1,15 @@
 # History
 
+
+## v0.1.X
+- Matrix equals スカラー値同士の比較時に誤差を指定できない問題を修正
+- Matrix isIdentity 正方行列以外でも正しく動作するように修正
+- Matrix isTridiagonal 正方行列の制限を撤去
+- Matrix 置換行列、上三角行列、下三角行列の判定を追加
+- LinearAlgebra norm1 列ベクトルの1ノルムの計算が誤っていたのを修正
+- LinearAlgebra rank 正方行列以外の行列で誤る場合があったのを修正
+
+
 ## v0.1.2
 - Jest 導入
 - Matrix 系以外についてテストを追加